refactor: Move toolAgent implementation to toolAgentCore.ts and update imports
- Move main toolAgent implementation from index.ts to toolAgentCore.ts\n- Update index.ts to re-export from toolAgentCore.ts\n- Remove deprecated toolAgent.ts file\n- Update all direct imports to use the new structure\n- Update documentation in README.md\n\nCloses #92
